+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Support CQL2 Array Functions
 - Ignored metadata fields can be configured in `fields.config.js`
 - PlayWright tests
+- Add config option `displayOverviewsForChildren` to toggle visualizing overviews for maps showing many STAC Items
 
 ### Changed